refactor: badge 컴포넌트를 tag 컴포넌트로 대체, 중복 컴포넌트 정리
